1_0_0_hotfix #bugfix within ventilation
The problem rose when the area and volume was not given to the _get_absolute_value_nominal in ventilation.py. In case the unit does not require volume and area, it should not calculate them within dictionary. a quick fix was done by considering the default value of area and volume to be 0. it is to note that the area and volume should be given if they are ought to be used in the calculation, so no problem is foreseen with this change. However, in case there is a problem with the calculation of ventilation nominal values, this commit is to be referred to.