first push, it is possible to do scenario creation
the scenarios are created as city objects. using the method in city called scenario creation. the init part of city is moved inside the new method called initalizer that is called everytime a simulate is called. the base scenarios should be simulated and then we would have mix and match later